# Optimise labels regex matchers containing a literal within the pattern