--- a/pkg/labels/regexp.go
+++ b/pkg/labels/regexp.go
@@ -23,6 +23,7 @@ type FastRegexMatcher struct {
 	re     *regexp.Regexp
 	prefix string
 	suffix string
+	contains string
 }
 
 func NewFastRegexMatcher(v string) (*FastRegexMatcher, error) {
@@ -41,7 +42,7 @@ func NewFastRegexMatcher(v string) (*FastRegexMatcher, error) {
 	}
 
 	if parsed.Op == syntax.OpConcat {
-		m.prefix, m.suffix = optimizeConcatRegex(parsed)
+		m.prefix, m.suffix, m.contains = optimizeConcatRegex(parsed)
 	}
 
 	return m, nil
@@ -54,6 +55,9 @@ func (m *FastRegexMatcher) MatchString(s string) bool {
 	if m.suffix != "" && !strings.HasSuffix(s, m.suffix) {
 		return false
 	}
+	if m.contains != "" && !strings.Contains(s, m.contains) {
+		return false
+	}
 	return m.re.MatchString(s)
 }
 
@@ -63,7 +67,7 @@ func (m *FastRegexMatcher) GetRegexString() string {
 
 // optimizeConcatRegex returns literal prefix/suffix text that can be safely
 // checked against the label value before running the regexp matcher.
-func optimizeConcatRegex(r *syntax.Regexp) (prefix, suffix string) {
+func optimizeConcatRegex(r *syntax.Regexp) (prefix, suffix, contains string) {
 	sub := r.Sub
 
 	// We can safely remove begin and end text matchers respectively
@@ -87,6 +91,16 @@ func optimizeConcatRegex(r *syntax.Regexp) (prefix, suffix string) {
 	}
 	if last := len(sub) - 1; sub[last].Op == syntax.OpLiteral {
 		suffix = string(sub[last].Rune)
+	}
+
+	// If contains any literal which is not a prefix/suffix, we keep the
+	// 1st one. We do not keep the whole list of literals to simplify the
+	// fast path.
+	for i := 1; i < len(sub) - 1; i++ {
+		if sub[i].Op == syntax.OpLiteral {
+			contains = string(sub[i].Rune)
+			break
+		}
 	}
 
 	return