@brian-brazil 2020-07-02
> BenchmarkPostingsForMatchers/Head/i=~".*1"-4                              7700854       10396696      +35.01%

It looks like suffixes are slower, which I expect would be a common thing.