@pracucci 2020-07-02
> > BenchmarkPostingsForMatchers/Head/i=~".*1"-4                              7700854       10396696      +35.01%
> 
> It looks like suffixes are slower, which I expect would be a common thing.

The main problem I have is that I can't get stable benchmark results on my laptop. I re-run the benchmark only for this specific case and the result is:

```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkPostingsForMatchers/Head/i=~".*1"-4      7194103       7180446       -0.19%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4     4634123       4872259       +5.14%

benchmark                                         old allocs     new allocs     delta
BenchmarkPostingsForMatchers/Head/i=~".*1"-4      6              6              +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4     7              7              +0.00%

benchmark                                         old bytes     new bytes     delta
BenchmarkPostingsForMatchers/Head/i=~".*1"-4      1605912       1605912       +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4     1605928       1605928       +0.00%
```