@brian-brazil 2020-07-02
Laptops tend to have CPU frequency throttling. Try turning that off.