@pracucci 2020-07-06
> Laptops tend to have CPU frequency throttling. Try turning that off.

Thanks for your feedback. I've managed to run the benchmark for the specific case that was spiky:

```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkPostingsForMatchers/Head/i=~".*1"-4      7144865       7135604       -0.13%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4     3706189       3684246       -0.59%

benchmark                                         old allocs     new allocs     delta
BenchmarkPostingsForMatchers/Head/i=~".*1"-4      6              6              +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4     7              7              +0.00%

benchmark                                         old bytes     new bytes     delta
BenchmarkPostingsForMatchers/Head/i=~".*1"-4      1605912       1605912       +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4     1605928       1605928       +0.00%
```