@pracucci 2020-07-06
I re-run the whole benchmark and now results look more sane (even if still suffering a bit of variance here and there). I think this is the best I can get out of my laptop. If you prefer, I can spin up a cloud instance and run benchmarks there.

```
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkPostingsForMatchers/Head/n="1"-4                                 797           795           -0.25%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-4                         1071          1056          -1.40%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-4                         1070          1064          -0.56%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-4                        1239          1208          -2.50%
BenchmarkPostingsForMatchers/Head/i=~".*"-4                               99792366      95160428      -4.64%
BenchmarkPostingsForMatchers/Head/i=~"1.*"-4                              21919168      22416796      +2.27%
BenchmarkPostingsForMatchers/Head/i=~".*1"-4                              7264742       6726004       -7.42%
BenchmarkPostingsForMatchers/Head/i=~".+"-4                               160866148     165400259     +2.82%
BenchmarkPostingsForMatchers/Head/i=~""-4                                 89026950      88507399      -0.58%
BenchmarkPostingsForMatchers/Head/i!=""-4                                 76772678      73130581      -4.74%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-4                 93975404      94659832      +0.73%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-4          94405392      95969937      +1.66%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-4                           74476178      72838562      -2.20%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-4                   72822329      72729016      -0.13%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-4                 176115921     161251128     -8.44%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-4                21923702      22007118      +0.38%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*1.*",j="foo"-4              165249786     102348121     -38.06%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-4          161813576     163480481     +1.03%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-4        184338225     193851524     +5.16%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~".*2.*",j="foo"-4      340250123     267306950     -21.44%
BenchmarkPostingsForMatchers/Block/n="1"-4                                34195         34239         +0.13%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-4                        592050        601196        +1.54%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-4                        601188        589446        -1.95%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-4                       596381        588014        -1.40%
BenchmarkPostingsForMatchers/Block/i=~".*"-4                              87874759      88325121      +0.51%
BenchmarkPostingsForMatchers/Block/i=~"1.*"-4                             14478893      14521794      +0.30%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4                             3638655       3612111       -0.73%
BenchmarkPostingsForMatchers/Block/i=~".+"-4                              102497642     102155124     -0.33%
BenchmarkPostingsForMatchers/Block/i=~""-4                                30036026      29980531      -0.18%
BenchmarkPostingsForMatchers/Block/i!=""-4                                21969562      21967733      -0.01%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-4                87547779      86514550      -1.18%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-4         87384553      87862680      +0.55%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-4                          21737308      21861386      +0.57%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-4                  23022924      22869153      -0.67%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-4                103254493     103824281     +0.55%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-4               14995961      15099611      +0.69%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*1.*",j="foo"-4             139737155     83194775      -40.46%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-4         102471406     103416985     +0.92%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-4       118245897     118461633     +0.18%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~".*2.*",j="foo"-4     241746617     185893392     -23.10%

benchmark                                                                 old allocs     new allocs     delta
BenchmarkPostingsForMatchers/Head/n="1"-4                                 6              6              +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-4                         11             11             +0.00%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-4                         11             11             +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-4                        12             12             +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*"-4                               9              9              +0.00%
BenchmarkPostingsForMatchers/Head/i=~"1.*"-4                              11140          11140          +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*1"-4                              6              6              +0.00%
BenchmarkPostingsForMatchers/Head/i=~".+"-4                               100039         100039         +0.00%
BenchmarkPostingsForMatchers/Head/i=~""-4                                 100043         100043         +0.00%
BenchmarkPostingsForMatchers/Head/i!=""-4                                 100039         100039         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-4                 14             14             +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-4          20             20             +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-4                           100044         100044         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-4                   100048         100048         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-4                 100048         100048         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-4                11149          11149          +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*1.*",j="foo"-4              40995          40995          +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-4          100054         100054         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-4        111250         111250         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~".*2.*",j="foo"-4      141100         141101         +0.00%
BenchmarkPostingsForMatchers/Block/n="1"-4                                6              6              +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-4                        11             11             +0.00%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-4                        11             11             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-4                       13             13             +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*"-4                              10             10             +0.00%
BenchmarkPostingsForMatchers/Block/i=~"1.*"-4                             11140          11140          +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4                             7              7              +0.00%
BenchmarkPostingsForMatchers/Block/i=~".+"-4                              100039         100039         +0.00%
BenchmarkPostingsForMatchers/Block/i=~""-4                                100043         100043         +0.00%
BenchmarkPostingsForMatchers/Block/i!=""-4                                100039         100039         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-4                15             15             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-4         21             21             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-4                          100044         100044         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-4                  100048         100048         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-4                100048         100048         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-4               11149          11149          +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*1.*",j="foo"-4             40995          40995          +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-4         100054         100054         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-4       111250         111250         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~".*2.*",j="foo"-4     141099         141100         +0.00%

benchmark                                                                 old bytes     new bytes     delta
BenchmarkPostingsForMatchers/Head/n="1"-4                                 296           296           +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-4                         424           424           +0.00%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-4                         424           424           +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-4                        456           456           +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*"-4                               1606012       1606011       -0.00%
BenchmarkPostingsForMatchers/Head/i=~"1.*"-4                              3147970       3147970       +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*1"-4                              1605912       1605912       +0.00%
BenchmarkPostingsForMatchers/Head/i=~".+"-4                               17264689      17264702      +0.00%
BenchmarkPostingsForMatchers/Head/i=~""-4                                 17264744      17264744      +0.00%
BenchmarkPostingsForMatchers/Head/i!=""-4                                 17264616      17264616      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-4                 1606132       1606139       +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-4          1606323       1606331       +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-4                           17264744      17264744      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-4                   17264872      17264872      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-4                 17264945      17264945      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-4                3148227       3148228       +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*1.*",j="foo"-4              7879683       7883364       +0.05%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-4          17265121      17265107      -0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-4        20417838      20417838      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~".*2.*",j="foo"-4      25150365      25150286      -0.00%
BenchmarkPostingsForMatchers/Block/n="1"-4                                296           296           +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-4                        424           424           +0.00%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-4                        424           424           +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-4                       1464          1464          +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*"-4                              1606027       1606019       -0.00%
BenchmarkPostingsForMatchers/Block/i=~"1.*"-4                             3147952       3147952       +0.00%
BenchmarkPostingsForMatchers/Block/i=~".*1"-4                             1605928       1605928       +0.00%
BenchmarkPostingsForMatchers/Block/i=~".+"-4                              17264645      17264645      +0.00%
BenchmarkPostingsForMatchers/Block/i=~""-4                                17264696      17264700      +0.00%
BenchmarkPostingsForMatchers/Block/i!=""-4                                17264600      17264600      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-4                1606152       1606159       +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-4         1606351       1606347       -0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-4                          17264728      17264728      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-4                  17264856      17264856      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-4                17264901      17264901      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-4               3148209       3148209       +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*1.*",j="foo"-4             7879659       7882274       +0.03%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-4         17265077      17265077      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-4       20417773      20417773      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~".*2.*",j="foo"-4     25150193      25150182      -0.00%
```